homework2/movie_theater_booking/requirements.txt
asgiref==3.8.1
Django==5.1.6
djangorestframework==3.15.2
gyp==0.1
pytz==2024.1
setuptools==68.1.2
six==1.16.0
sqlparse==0.5.3
supervisor==4.2.5
wheel==0.42.0
